backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pymongo==4.6.1
python-dotenv==1.0.1
pydantic[email]==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
reportlab==4.0.9
